# cmd/compile: new inline heuristic for struct compares